Clean up the scanner
