+++ scanner/requirements.txt
@@ -2,3 +2,4 @@ guessit
 aiohttp
 jsons
 black-with-tabs
+watchdog
